mf_collab_client::yrs

Struct XmlTextRef

pub struct XmlTextRef(/* private fields */);

A shared data type used for collaborative text editing, that can be used in a context of XmlElementRef node. It enables multiple users to add and remove chunks of text in efficient manner. This type is internally represented as a mutable double-linked list of text chunks

an optimization occurs during [Transaction::commit], which allows to squash multiple consecutively inserted characters together as a single chunk of text even between transaction boundaries in order to preserve more efficient memory model.

Just like XmlElementRef, XmlTextRef can be marked with extra metadata in form of attributes.

XmlTextRef structure internally uses UTF-8 encoding and its length is described in a number of bytes rather than individual characters (a single UTF-8 code point can consist of many bytes).

Like all Yrs shared data types, XmlTextRef is resistant to the problem of interleaving (situation when characters inserted one after another may interleave with other peers concurrent inserts after merging all updates together). In case of Yrs conflict resolution is solved by using unique document id to determine correct and consistent ordering.

XmlTextRef offers a rich text editing capabilities (it’s not limited to simple text operations). Actions like embedding objects, binaries (eg. images) and formatting attributes are all possible using XmlTextRef.

Keep in mind that XmlTextRef::get_string method returns a raw string, while rendering formatting attrbitues as XML tags in-text. However it doesn’t include embedded elements. If there’s a need to include them, use XmlTextRef::diff method instead.

Another note worth reminding is that human-readable numeric indexes are not good for maintaining cursor positions in rich text documents with real-time collaborative capabilities. In such cases any concurrent update incoming and applied from the remote peer may change the order of elements in current XmlTextRef, invalidating numeric index. For such cases you can take advantage of fact that XmlTextRef implements IndexedSequence::sticky_index method that returns a permanent index position that sticks to the same place even when concurrent updates are being made.

Example

use yrs::{Any, Array, ArrayPrelim, Doc, GetString, Text, Transact, WriteTxn, XmlFragment, XmlTextPrelim};
use yrs::types::Attrs;

let doc = Doc::new();
let mut txn = doc.transact_mut();
let f = txn.get_or_insert_xml_fragment("article");
let text = f.insert(&mut txn, 0, XmlTextPrelim::new(""));

let bold = Attrs::from([("b".into(), true.into())]);
let italic = Attrs::from([("i".into(), true.into())]);

text.insert(&mut txn, 0, "hello ");
text.insert_with_attributes(&mut txn, 6, "world", italic);
text.format(&mut txn, 0, 5, bold);

assert_eq!(text.get_string(&txn), "<b>hello</b> <i>world</i>");

// remove formatting
let remove_italic = Attrs::from([("i".into(), Any::Null)]);
text.format(&mut txn, 6, 5, remove_italic);

assert_eq!(text.get_string(&txn), "<b>hello</b> world");

// insert binary payload eg. images
let image = b"deadbeaf".to_vec();
text.insert_embed(&mut txn, 1, image);

// insert nested shared type eg. table as ArrayRef of ArrayRefs
let table = text.insert_embed(&mut txn, 5, ArrayPrelim::default());
let header = table.insert(&mut txn, 0, ArrayPrelim::from(["Book title", "Author"]));
let row = table.insert(&mut txn, 1, ArrayPrelim::from(["\"Moby-Dick\"", "Herman Melville"]));