mf_collab_client::yrs

Module sync

Modules

awareness

protocol

Structs

Awareness

The Awareness class implements a simple shared state protocol that can be used for non-persistent data like awareness information (cursor, username, status, ..). Each client can update its own local state and listen to state changes of remote clients.

AwarenessUpdate

A structure that represents an encodable state of an Awareness struct.

DefaultProtocol

A default implementation of y-sync Protocol.

MessageReader

Since y-sync protocol enables for a multiple messages to be packed into a singe byte payload, MessageReader can be used over the decoder to read these messages one by one in iterable fashion.

Enums

Error

An error type returned in response from y-sync Protocol.

Message

SyncMessage

Traits

Protocol

Trait implementing a y-sync protocol. The default implementation can be found in DefaultProtocol, but its implementation steps can be potentially changed by the user if necessary.