mf_collab_client::yrs

Trait DeepObservable

pub trait DeepObservable: AsRef<Branch> {
    // Provided method
    fn observe_deep<F>(&self, f: F) -> Subscription
       where F: Fn(&TransactionMut<'_>, &Events) + 'static { ... }
}

Trait implemented by all Y-types, allowing for observing events which are emitted by nested types.

Provided Methods

fn observe_deep<F>(&self, f: F) -> Subscription
where F: Fn(&TransactionMut<'_>, &Events) + 'static,

Subscribe a callback f for all events emitted by this and nested collaborative types. Callback is accepting transaction which triggered that event and event itself, wrapped within an Event structure.

In case when a nested shared type (e.g. MapRef,ArrayRef,TextRef) is being removed, all of its contents will be removed first. So the observed value will be empty. For example, The value wrapped in the EntryChange::Removed of the Event::Map will be empty.

This method returns a subscription, which will automatically unsubscribe current callback when dropped.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.