mf_collab_client::yrs

Trait IndexedSequence

pub trait IndexedSequence: AsRef<Branch> {
    // Provided method
    fn sticky_index(
        &self,
        txn: &mut TransactionMut<'_>,
        index: u32,
        assoc: Assoc,
    ) -> Option<StickyIndex> { ... }
}

Trait used to retrieve a StickyIndex corresponding to a given human-readable index. Unlike standard indexes StickyIndex enables to track the location inside of a shared y-types, even in the face of concurrent updates.

Provided Methods

fn sticky_index( &self, txn: &mut TransactionMut<'_>, index: u32, assoc: Assoc, ) -> Option<StickyIndex>

Returns a StickyIndex equivalent to a human-readable index. Returns None if index is beyond the length of current sequence.