mf_collab_client::yrs

Trait Xml

pub trait Xml: AsRef<Branch> {
    // Provided methods
    fn parent(&self) -> Option<XmlNode> { ... }
    fn remove_attribute<K>(&self, txn: &mut TransactionMut<'_>, attr_name: &K)
       where K: AsRef<str> { ... }
    fn insert_attribute<K, V>(
        &self,
        txn: &mut TransactionMut<'_>,
        attr_name: K,
        attr_value: V,
    )
       where K: Into<Arc<str>>,
             V: Into<String> { ... }
    fn get_attribute<T>(&self, txn: &T, attr_name: &str) -> Option<String>
       where T: ReadTxn { ... }
    fn attributes<'a, T>(&'a self, txn: &'a T) -> Attributes<'a, &'a T, T> 
       where T: ReadTxn { ... }
    fn siblings<'a, T>(&self, txn: &'a T) -> Siblings<'a, T> 
       where T: ReadTxn { ... }
}

Provided Methods

fn parent(&self) -> Option<XmlNode>

fn remove_attribute<K>(&self, txn: &mut TransactionMut<'_>, attr_name: &K)
where K: AsRef<str>,

Removes an attribute recognized by an attr_name from a current XML element.

fn insert_attribute<K, V>( &self, txn: &mut TransactionMut<'_>, attr_name: K, attr_value: V, )
where K: Into<Arc<str>>, V: Into<String>,

Inserts an attribute entry into current XML element.

fn get_attribute<T>(&self, txn: &T, attr_name: &str) -> Option<String>
where T: ReadTxn,

Returns a value of an attribute given its attr_name. Returns None if no such attribute can be found inside of a current XML element.

fn attributes<'a, T>(&'a self, txn: &'a T) -> Attributes<'a, &'a T, T>
where T: ReadTxn,

Returns an unordered iterator over all attributes (key-value pairs), that can be found inside of a current XML element.

fn siblings<'a, T>(&self, txn: &'a T) -> Siblings<'a, T>
where T: ReadTxn,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.