mf_collab_client::yrs::types

Module array

Structs

ArrayEvent

Event generated by ArrayRef::observe method. Emitted during transaction commit phase.

ArrayIter

ArrayPrelim

A preliminary array. It’s can be used to initialize an YArray, when it’s about to be nested into another Yrs data collection, such as [Map] or another YArray.

ArrayRef

A collection used to store data in an indexed sequence structure. This type is internally implemented as a double linked list, which may squash values inserted directly one after another into single list node upon transaction commit.

Traits

Array