mf_collab_client::yrs::types

Module map

Structs

Keys

An unordered iterator over the keys of a Map.

MapEvent

Event generated by [Map::observe] method. Emitted during transaction commit phase.

MapIter

MapPrelim

A preliminary map. It can be used to early initialize the contents of a Map, when it’s about to be inserted into another Yrs collection, such as Array or another Map.

MapRef

Collection used to store key-value entries in an unordered manner. Keys are always represented as UTF-8 strings. Values can be any value type supported by Yrs: JSON-like primitives as well as shared data types.

Values

Iterator over the values of a Map.

Traits

Map