mf_collab_client::yrs::types

Trait Map

pub trait Map: Sized + AsRef<Branch> {
    // Provided methods
    fn len<T>(&self, _txn: &T) -> u32
       where T: ReadTxn { ... }
    fn keys<'a, T>(&'a self, txn: &'a T) -> Keys<'a, &'a T, T> 
       where T: ReadTxn + 'a { ... }
    fn values<'a, T>(&'a self, txn: &'a T) -> Values<'a, &'a T, T> 
       where T: ReadTxn + 'a { ... }
    fn iter<'a, T>(&'a self, txn: &'a T) -> MapIter<'a, &'a T, T> 
       where T: ReadTxn + 'a { ... }
    fn insert<K, V>(
        &self,
        txn: &mut TransactionMut<'_>,
        key: K,
        value: V,
    ) -> <V as Prelim>::Return
       where K: Into<Arc<str>>,
             V: Prelim { ... }
    fn remove(&self, txn: &mut TransactionMut<'_>, key: &str) -> Option<Value> { ... }
    fn get<T>(&self, txn: &T, key: &str) -> Option<Value>
       where T: ReadTxn { ... }
    fn contains_key<T>(&self, _txn: &T, key: &str) -> bool
       where T: ReadTxn { ... }
    fn clear(&self, txn: &mut TransactionMut<'_>) { ... }
}

Provided Methods

fn len<T>(&self, _txn: &T) -> u32
where T: ReadTxn,

Returns a number of entries stored within current map.

fn keys<'a, T>(&'a self, txn: &'a T) -> Keys<'a, &'a T, T>
where T: ReadTxn + 'a,

Returns an iterator that enables to traverse over all keys of entries stored within current map. These keys are not ordered.

fn values<'a, T>(&'a self, txn: &'a T) -> Values<'a, &'a T, T>
where T: ReadTxn + 'a,

Returns an iterator that enables to traverse over all values stored within current map.

fn iter<'a, T>(&'a self, txn: &'a T) -> MapIter<'a, &'a T, T>
where T: ReadTxn + 'a,

Returns an iterator that enables to traverse over all entries - tuple of key-value pairs - stored within current map.

fn insert<K, V>( &self, txn: &mut TransactionMut<'_>, key: K, value: V, ) -> <V as Prelim>::Return
where K: Into<Arc<str>>, V: Prelim,

Inserts a new value under given key into current map. Returns an integrated value.

fn remove(&self, txn: &mut TransactionMut<'_>, key: &str) -> Option<Value>

Removes a stored within current map under a given key. Returns that value or None if no entry with a given key was present in current map.

Removing nested shared types

In case when a nested shared type (eg. MapRef, ArrayRef, TextRef) is being removed, all of its contents will also be deleted recursively. A returned value will contain a reference to a current removed shared type (which will be empty due to all of its elements being deleted), not the content prior the removal.

fn get<T>(&self, txn: &T, key: &str) -> Option<Value>
where T: ReadTxn,

Returns a value stored under a given key within current map, or None if no entry with such key existed.

fn contains_key<T>(&self, _txn: &T, key: &str) -> bool
where T: ReadTxn,

Checks if an entry with given key can be found within current map.

fn clear(&self, txn: &mut TransactionMut<'_>)

Clears the contents of current map, effectively removing all of its entries.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.