mf_collab_client::yrs::types

Module xml

Structs

Attributes

Iterator over the attributes (key-value pairs represented as a strings) of an [XmlElement].

Siblings

TreeWalker

An iterator over [XmlElement] successors, working in a recursive depth-first manner.

XmlElementPrelim

A preliminary type that will be materialized into an XmlElementRef once it will be integrated into Yrs document.

XmlElementRef

XML element data type. It represents an XML node, which can contain key-value attributes (interpreted as strings) as well as other nested XML elements or rich text (represented by XmlTextRef type).

XmlEvent

Event generated by [XmlElement::observe] method. Emitted during transaction commit phase.

XmlFragmentPrelim

A preliminary type that will be materialized into an XmlFragmentRef once it will be integrated into Yrs document.

XmlFragmentRef

A XML fragment, which works as an untagged collection of XML nodes.

XmlHookRef

(Obsolete) an Yjs-compatible XML node used for nesting Map elements.

XmlTextEvent

Event generated by [XmlText::observe] method. Emitted during transaction commit phase.

XmlTextPrelim

A preliminary type that will be materialized into an XmlTextRef once it will be integrated into Yrs document.

XmlTextRef

A shared data type used for collaborative text editing, that can be used in a context of XmlElementRef node. It enables multiple users to add and remove chunks of text in efficient manner. This type is internally represented as a mutable double-linked list of text chunks

Enums

XmlNode

An return type from XML elements retrieval methods. It’s an enum of all supported values, that can be nested inside of XmlElementRef. These are other XmlElementRefs, XmlFragmentRefs or XmlTextRef values.

Traits

Xml

XmlFragment

XmlPrelim

Trait shared by preliminary types that can be used as XML nodes: XmlElementPrelim, XmlFragmentPrelim and XmlTextPrelim.