mf_collab_client::yrs

Module undo

Structs

Event

Options

Set of options used to configure UndoManager.

StackItem

A unit of work for the UndoManager. It contains a compressed information about all updates and deletions tracked by a corresponding undo manager. Whenever an [UndoManger::undo] or UndoManager::redo methods are called a last StackItem is being used to modify a state of the document.

UndoManager

Undo manager is a structure used to perform undo/redo operations over the associated shared type(s).

Enums

EventKind

Enum which informs if correlated Event was produced as a result of either undo or redo operation over UndoManager.